src/edfi-paging-test/pyproject.toml
[tool.poetry]
name = "edfi-paging-test"
version = "0.1.0"
description = "Ed-Fi Data Out Test"
authors = ["Ed-Fi Alliance, LLC and contributors"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
ConfigArgParse = "^1.5.3"
python-dotenv = "^0.19.2"
requests-oauthlib = "^1.3.1"
opnieuw = "^1.1.0"
requests = "^2.27.1"
types-requests = "^2.27.11"
urllib3 = "^1.26.8"
requests-mock = "^1.9.3"
pytest-mock = "^3.7.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^7.0.1"
mypy = "^0.931"
pytest-describe = "^2.0.1"
coverage = "^6.3.2"
black = "^22.1.0"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]